+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - `<Table />` - hide pagination when filtered & page size can show all filtered.
+- `<Table />` - pagination count number of possibly filterd rows.
 
 ## [1.2.0] - 2020-06-30
 ### Added