+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `<Carousel/>` - component added with examples.
+- `<ImageSlider/>` - component added with examples.
 
 ## [1.3.0] - 2020-07-06
 ### Fixed
@@ -17,8 +20,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - `<StyledCardContainer/>` - component added with examples.
-- `<Carousel/>` - component added with examples.
-- `<ImageSlider/>` - component added with examples.
 
 ## [1.2.0] - 2020-06-30
 ### Added